+++ routes/reviews.js
@@ -1,0 +1,35 @@
+const express = require('express');
+const router = express.Router( {mergeParams: true }); 
+const Campground = require('../models/campground');
+const Review = require('../models/review');
+const ExpressError = require('../utils/ExpressError')
+const catchAsync = require('../utils/catchAsync');
+const { reviewSchema } = require('../schema');
+
+const validateReview = (req ,res, next) => {
+    const { error } = reviewSchema.validate(req.body);
+    if(error){
+        const msg = error.details.map(err => err.message).join(',')
+        throw new ExpressError (msg, 404)
+    }else{
+        next()
+    }
+}
+
+router.delete('/:reviewId', catchAsync(async(req, res) => {
+    const { id, reviewId } = req.params;
+    await Campground.findByIdAndUpdate(id, {$pull: {reviews: reviewId}});
+    await Review.findByIdAndDelete(reviewId)
+    res.redirect(`/campgrounds/${id}`)
+}))
+
+router.post('/', validateReview, catchAsync(async (req, res) => {
+    const campground = await Campground.findById(req.params.id);
+    const review = new Review(req.body.review);
+    campground.reviews.push(review);
+    await review.save();
+    await campground.save();
+    res.redirect(`/campgrounds/${campground._id}`)
+}))
+
+module.exports = router;
